% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.sparsegl-methods.R
\name{coef.cv.sparsegl}
\alias{coef.cv.sparsegl}
\title{Extract coefficients from a `cv.sparsegl` object.}
\usage{
\method{coef}{cv.sparsegl}(object, s = c("lambda.1se", "lambda.min"), ...)
}
\arguments{
\item{object}{Fitted [cv.sparsegl()] object.}

\item{s}{Value(s) of the penalty parameter `lambda` at which
coefficients are desired. Default is the single
value `s = "lambda.1se"` stored in the CV object (corresponding to
the largest value of `lambda` such that CV error estimate is within 1
standard error of the minimum). Alternatively `s = "lambda.min"` can be
used (corresponding to the minimum of cross validation error estimate).
If `s` is numeric, it is taken as the value(s) of `lambda` to be used.}

\item{...}{Not used.}
}
\value{
The coefficients at the requested value(s) for `lambda`.
}
\description{
This function etracts coefficients from a
cross-validated [sparsegl()] model, using the stored `"sparsegl.fit"`
object, and the optimal value chosen for `lambda`.
}
\seealso{
[cv.sparsegl()] and [predict.cv.sparsegl()].
}
