% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_risk.R
\name{estimate_risk}
\alias{estimate_risk}
\title{Calculate information criteria.}
\usage{
estimate_risk(object, x, type = c("AIC", "BIC", "GCV"), approx_df = FALSE)
}
\arguments{
\item{object}{fitted object from a call to [sparsegl()].}

\item{x}{Matrix. The matrix of predictors used to estimate
the `sparsegl` object. May be missing if `approx_df = TRUE`.}

\item{type}{one or more of AIC, BIC, or GCV.}

\item{approx_df}{the `df` component of a `sparsegl` object is an
approximation (albeit a fairly accurate one) to the actual degrees-of-freedom.
However, the exact value requires inverting a portion of `X'X`. So this
computation may take some time (the default computes the exact df).}
}
\value{
a `data.frame` with as many rows as `object$lambda`. It contains
  columns `lambda`, `df`, and the requested risk types.
}
\description{
This function uses the degrees of freedom to calculate various information
criteria. This function uses the "unknown variance" version of the likelihood.
Only implemented for Gaussian regression. The constant is ignored (as in
[stats::extractAIC()]).
}
\references{
Vaiter S, Deledalle C, Peyré G, Fadili J, Dossal C. (2012). \emph{The
Degrees of Freedom of the Group Lasso for a General Design}.
\url{https://arxiv.org/pdf/1212.6478.pdf}.
}
\seealso{
[sparsegl()] method.
}
