% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Example dataset}
\format{
A data frame with X rows and Y columns:
\describe{
  \item{X}{Matrix of predictors}
  \item{Y}{Outcome variable}
  \item{hier_info}{Group information passed to hierNest functions}
}
}
\usage{
data(example_data)
}
\description{
Example dataset for the sample code of hierNest
}
\keyword{datasets}
