% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cv.sparsegl.R
\name{plot.cv.sparsegl}
\alias{plot.cv.sparsegl}
\title{Plot cross-validation curves produced from a `cv.sparsegl` object.}
\usage{
\method{plot}{cv.sparsegl}(x, log_axis = c("xy", "x", "y", "none"), sign.lambda = 1, ...)
}
\arguments{
\item{x}{Fitted `"cv.sparsegl"` object, produced with [cv.sparsegl()].}

\item{log_axis}{Apply log scaling to the requested axes.}

\item{sign.lambda}{Either plot against `log(lambda)` (default) or the
reverse if `sign.lambda < 0`.}

\item{...}{Not used.}
}
\description{
Plots the cross-validation curve, and upper and lower standard deviation
curves, as a function of the `lambda` values used.
}
\details{
A [ggplot2::ggplot()] plot is produced. Additional user
modifications may be added as desired.
}
\seealso{
[cv.sparsegl()].
}
