% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.sparsegl.R
\name{plot.sparsegl}
\alias{plot.sparsegl}
\title{Plot solution paths from a `sparsegl` object.}
\usage{
\method{plot}{sparsegl}(
  x,
  y_axis = c("coef", "group"),
  x_axis = c("lambda", "penalty"),
  add_legend = n_legend_values < 20,
  ...
)
}
\arguments{
\item{x}{Fitted `"sparsegl"` object, produced by [sparsegl()].}

\item{y_axis}{Variable on the y_axis. Either the coefficients (default)
or the group norm.}

\item{x_axis}{Variable on the x-axis. Either the (log)-lambda
sequence (default) or the value of the penalty. In the second case,
the penalty is scaled by its maximum along the path.}

\item{add_legend}{Show the legend. Often, with many groups/predictors, this
can become overwhelming. The default produces a legend if the number of
groups/predictors is less than 20.}

\item{...}{Not used.}
}
\description{
Produces a coefficient profile plot of a fitted
[sparsegl()] object. The result is a [ggplot2::ggplot()]. Additional user
modifications can be added as desired.
}
\seealso{
[sparsegl()].
}
